solucionando...

+++ AllersProject/Modelo/AssociationAnalyzer.cs
@@ -113,7 +113,7 @@ namespace Modelo
                     }
                 }
             }
-            double minimunSupport = binaryTransactions.Count * minSupport;
+            int minimunSupport =(int)( binaryTransactions.Count * minSupport);
             for(int i = 0; i < candidateSet.Count; i++)
             {
                 if (itemSetToSupport[candidateSet[i]] <minimunSupport)
